requirements.txt
appdirs==1.4.4
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
cycler==0.10.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.3.0
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
Flask==1.1.2
geojson==2.5.0
gmplot==1.2.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.1
importlib-resources==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
nautical-calculations==1.0.1
numpy==1.18.5
pandas==1.0.4
psycopg2==2.8.5
PyJWT==1.7.1
pyparsing==2.4.7
pyproj==2.6.1.post1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
virtualenv==20.0.21
Werkzeug==1.0.1
whitenoise==5.1.0
zipp==3.1.0
